Created Ratings folder & ratingStar component,also updated main & popular components with the ratingStar
